requirements.txt
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
matplotlib>=3.7.0
openai>=1.0.0
